requirements.txt
schedule
art
python-dotenv
pytz
rich